+++ BE/application/src/main/java/com/travel/application/travelservice/createstatustrip/service/IRepoCreateStatusTrip.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.travel.application.entity.StatusTripEntity;
 
-public interface IRepoCreateStatusTrip  extends JpaRepository<StatusTripEntity, Integer>{
+public interface IRepoCreateStatusTrip  extends JpaRepository<StatusTripEntity, Long>{
 
 }
