+++ BE/application/src/main/java/com/travel/application/travelservice/getroute/service/IRepoGetRoute.java
@@ -8,7 +8,7 @@ import com.travel.application.entity.ProvinceEntity;
 import com.travel.application.entity.RouteEntity;
 
 @Repository
-public interface IRepoGetRoute extends JpaRepository<RouteEntity, Integer> {
+public interface IRepoGetRoute extends JpaRepository<RouteEntity, Long> {
 		// Use Query
 //	@Query(value="select * from route where province_start = ?1", nativeQuery = true)
 //	public RouteEntity findRouteByProvinceStart(int provinceStart);
