+++ BE/application/src/main/java/com/travel/application/travelservice/getstatustrip/service/IRepoGetStatusTrip.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.travel.application.entity.StatusTripEntity;
 
-public interface IRepoGetStatusTrip  extends JpaRepository<StatusTripEntity, Integer>{
+public interface IRepoGetStatusTrip  extends JpaRepository<StatusTripEntity, Long>{
 
 }
